requirements.txt
fastapi
uvicorn
pinecone-client
python-dotenv
scikit-learn
openai
